RFC: support overriding String with reflection
If a value implements the UseReflection interface by providing a
LogWithReflection method, then the normal handling (in particular, using the
Stringer interface if available) is overridden and the value gets
logged via reflection.

This is useful for klog.ObjectRef which provides a String method for logging as
plain text, but is meant to be logged as struct when using JSON output.