implement flushing on demand
Flushing through a method provided by the zapr LogSink makes it easier for
consumers who only have access to the abstract logr.Logger to flush buffered
messages. Other LogSinks might implement the same interface.

The alternative is to go through Underlier, but that is more complicated and
more tightly couples the caller with details of the LogSink.