support "structured error"
A "structured error" is an error which implements error and `ErrorDetails() any`.
When encountering such an error, it gets logged normally and then another field
gets added with the result of ErrorDetails.

The name of the extra field is the original field name plus a configurable
suffix, with "Details" as default.

The extra details are logged as if they had been passed as a value to
zapr. There's no protection against infinite recursion when the additional
details are the original error.

A simple guard against this (not expanding error again while formatting an
error) is too simplistic and would prevent nice rendering of a wrapped error
that might get returned by MarshalLog. zapIt would have to track the exact
error instance and detect when it gets the same value again.