support slog + logr 1.3.0
When built with Go >= 1.21, zapr implements an additional interface which
adds support for directly logging a slog.Record. The verbosity level in
such records gets adjusted by the logger's verbosity, but only if the record
has a level < slog.LevelError.

To use zapr as slog handler, use slogr.NewSlogHandler(zapr.NewLogger(...)).

In addition to supporting usage as a SlogHandler, special slog values (Group,
LogValuer) are also supported, regardless of which front-end API is used.