update linter config and fix issues
Selecting linters in the workflow file has the disadvantage that a manual
"golangci-lint run" invocation uses different settings. It's better to do it in
a configuration file.

Several of the selected linters were deprecated and thus get removed:

    WARN [runner] The linter 'structcheck' is deprecated (since v1.49.0) due to: The owner seems to have abandoned the linter. Replaced by unused.
    WARN [runner] The linter 'deadcode' is deprecated (since v1.49.0) due to: The owner seems to have abandoned the linter. Replaced by unused.
    WARN [runner] The linter 'varcheck' is deprecated (since v1.49.0) due to: The owner seems to have abandoned the linter. Replaced by unused.

Some issues were reported with this configuration by golangci-lint 1.52.2 and get fixed.